Changed objectId type to size_t

+++ include/MainWindow.h
@@ -94,7 +94,7 @@ public slots:
     bool saveFileDefaultPathId(int documentId);
     void onActiveDocumentChanged(int newIndex);
     void tabCloseRequested(int i) ;
-    void objectTreeWidgetSelectionChanged(unsigned int objectId);
+    void objectTreeWidgetSelectionChanged(size_t objectId);
     void newFile(); // empty new file
     void openFile(const QString& filePath);
     void updateMouseButtonObjectState();
